pyproject.toml
[tool.poetry]
name = "spherinator"
version = "0.3.0"
description = "Rotation-Invariant Hyperspherical Variational Convolutional Autoencoder"
authors = [
    "Kai Polsterer <kai.polsterer@h-its.org>",
    "Bernd Doser <bernd.doser@h-its.org>",
    "Andreas Fehlner <andreas.fehlner@h-its.org>",
    "Sebastian T. Gomez <sebastian.trujillogomez@h-its.org>",
]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/HITS-AIN/Spherinator"
packages = [
    { include = "spherinator", from = "src" },
    { include = "hipster", from = "src" },
]
include = [
    { path = "tests", format = "sdist" },
]

[tool.poetry.scripts]
hipster = 'hipster.main:main'
spherinator = 'spherinator.main:main'

[tool.poetry.dependencies]
python = ">=3.9,<3.13"

astropy = "^6.0.0"
healpy = "^1.16.6"
ipykernel = "^6.27.1"
jsonargparse = {extras = ["signatures"], version = "^4.27.1"}
lightning = "^2.1.2"
matplotlib = "^3.8.2"
numpy = "^1.26.2"
onnx = "^1.15.0"
onnxscript = "0.1.0.dev20240626"
pandas = "^2.1.3"
power-spherical = "^0.8.1"
psutil = "^5.9.7"
scikit-image = "^0.22.0"
scipy = "^1.11.4"

torch = "^2.2"
# torch = {version = "^2.1.2", source = "pytorch-nightly-cu121"}
# torch = {url = "https://download.pytorch.org/whl/nightly/cu121/torch-2.3.0.dev20240101%2Bcu121-cp39-cp39-linux_x86_64.whl"}

torchvision = "^0"
# torchvision = {version = "0.18.0.dev20231221", source = "pytorch-nightly-cu121"}
# torchvision = {url = "https://download.pytorch.org/whl/nightly/cu121/torchvision-0.18.0.dev20240101%2Bcu121-cp39-cp39-linux_x86_64.whl"}
# pytorch-triton = {url = "https://download.pytorch.org/whl/nightly/pytorch_triton-2.2.0%2Be28a256d71-cp39-cp39-linux_x86_64.whl"}

flytekit = "^1.11.0"
scikit-learn = "^1.3.2"
flytekitplugins-kfpytorch = "^1.12.2"
flytekitplugins-envd = "^1.12.2"
torchdata = "^0.7.1"
kubernetes = "^30.1.0"
requests = "2.31.0"
plotly = "^5.22.0"
flytekitplugins-deck-standard = "^1.13.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
jupytext = "^1.15.2"
pip-licenses = "^4.3.3"
pytest = "^8.0.2"
pytest-cov = "^5.0"
tensorboard = "^2.15.1"
wandb = "^0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
myst-parser = "^3.0.1"
recommonmark = "^0.7.1"
sphinx = "^7.3.7"
sphinx-autodoc-typehints = "^2.1.0"
sphinx-rtd-theme = "^2.0.0"

[[tool.poetry.source]]
name = "pypi"
priority = "primary"

[[tool.poetry.source]]
name = "pytorch-nightly-cu121"
url = "https://download.pytorch.org/whl/nightly/cu121"
priority = "explicit"

[[tool.poetry.source]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest]
timeout = 300
testpaths = 'tests'
xfail_strict = true

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore:The 'train_dataloader' does not have many workers which may be a bottleneck. Consider increasing the value of the `num_workers` argument",
    "ignore:torch.onnx.dynamo_export only implements opset version 18 for now.",
    "ignore:GPU available but not used.",
    "ignore:You requested to overfit but enabled train dataloader shuffling.",
]

[tool.pylint.TYPECHECK]
generated-members = 'numpy.*, torch.*'
